+++ kernel/src/ipc/kill.rs
@@ -1,7 +1,7 @@
-use crate::arch::ipc::signal::{SigCode, Signal};
 use crate::ipc::signal_types::{SigInfo, SigType};
 use crate::process::pid::{Pid, PidType};
 use crate::process::{ProcessControlBlock, ProcessManager, RawPid};
+use crate::{arch::ipc::signal::Signal, ipc::signal_types::SigCode};
 use alloc::sync::Arc;
 use core::sync::atomic::compiler_fence;
 use system_error::SystemError;
