+++ kernel/src/process/timer.rs
@@ -1,10 +1,10 @@
-use crate::arch::ipc::signal::{SigCode, Signal};
 use crate::exception::InterruptArch;
 use crate::ipc::signal_types::SigType;
 use crate::process::CurrentIrqArch;
 use crate::process::RawPid;
 use crate::process::SigInfo;
 use crate::time::timer::{clock, Jiffies, Timer, TimerFunction};
+use crate::{arch::ipc::signal::Signal, ipc::signal_types::SigCode};
 use alloc::{boxed::Box, sync::Arc};
 use core::sync::atomic::compiler_fence;
 use core::time::Duration;
